// src/models/IVehicleEventModel.ts
export interface IVehicleEventModel {
  eventNumber?: number;
  vehicleNumber?: number;
  vehicleEventNumber?: number;
  vehicleNumberThisVehicle?: number;
  areaOfImpactThisVehicle?: number;
  areaOfImpactThisVehicleName?: string;
  sequenceOfEvents?: number;
  sequenceOfEventsName?: string;
  vehicleNumberOtherVehicle?: number;
  areaOfImpactOtherVehicle?: number;
  areaOfImpactOtherVehicleName?: string;
}